Move makeParallel() from commands/index.ts to commands/helper/parallel.ts
