+++ src/commands/helpers/parallel.ts
@@ -1,0 +1,3 @@
+export function makeParallel<T>(concurrency: number, promiseFactory: () => Thenable<T>): Thenable<T[]> {
+  return Promise.all(Array.from({ length: concurrency }, promiseFactory));
+}
